requirements.txt
pip
rpi_ws281x
adafruit-circuitpython-neopixel
board
neopixel
requests
datetime